Rename service, router, repository의 users -> user
