+++ routes/user.js
@@ -1,9 +1,9 @@
 const express = require("express");
 const router = express.Router();
 const { userJoinValidator } = require("../middlewares/validator/userValidator");
-const { usersService } = require("../services/index");
+const { userService } = require("../services/index");
 
 // user create에 대한 api
-router.post("/join", userJoinValidator(), usersService.addUser);
+router.post("/join", userJoinValidator(), userService.addUser);
 
 module.exports = router;
